refactor: NlpToken 위치 이동
